Intent: Reorganize configuration coherence
Operators involved: [Self-organization, Coupling]

Affected invariants: [#7, #8]

Key changes:\n- Create tnfr.config package with init.py, constants.py, presets.py\n- Update engine and tests to import glyph constants and presets from tnfr.config\n- Add compatibility shims for tnfr.constants_glyphs and tnfr.presets with deprecation warnings

Expected risks/dissonances: Deprecation warnings for legacy imports may surface in downstream consumers; validated that shims maintain API.

Metrics: C(t) unchanged; νf unaffected; no phase alterations expected.

Equivalence map: tnfr.config.load_config -> tnfr.config.init.load_config; tnfr.constants_glyphs -> tnfr.config.constants; tnfr.presets -> tnfr.config.presets
