+++ tests/test_observers.py
@@ -17,7 +17,7 @@ from tnfr.observers import (
 )
 from tnfr.gamma import kuramoto_R_psi
 from tnfr.sense import sigma_vector
-from tnfr.constants_glyphs import ANGLE_MAP
+from tnfr.config.constants import ANGLE_MAP
 from tnfr.helpers.numeric import angle_diff
 from tnfr.alias import set_attr
 from tnfr.callback_utils import CallbackEvent
